+++ dbx/api/launch/runners/asset_based.py
@@ -5,8 +5,9 @@ from typing import Optional, Union, Tuple, Dict, Any
 from databricks_cli.sdk import ApiClient, JobsService
 
 from dbx.api.configure import ProjectConfigurationManager
-from dbx.api.launch.functions import load_dbx_file
 from dbx.api.launch.processors import ClusterReusePreprocessor
+from dbx.api.launch.runners.base import RunData
+from dbx.api.storage.io import StorageIO
 from dbx.models.deployment import EnvironmentDeploymentInfo
 from dbx.models.workflow.v2dot0.parameters import AssetBasedRunPayload as V2dot0AssetBasedParametersPayload
 from dbx.models.workflow.v2dot1.parameters import AssetBasedRunPayload as V2dot1AssetBasedParametersPayload
@@ -39,14 +40,14 @@ class AssetBasedLauncher:
         else:
             return V2dot1AssetBasedParametersPayload.from_string(payload)
 
-    def launch(self) -> Tuple[Dict[Any, Any], Optional[str]]:
+    def launch(self) -> Tuple[RunData, Optional[int]]:
         dbx_echo(
             f"Launching workflow in assets-based mode "
             f"(via RunSubmit method, Jobs API V{self.api_client.jobs_api_version})"
         )
 
         service = JobsService(self.api_client)
-        env_spec = load_dbx_file(self.run_id, "deployment-result.json")
+        env_spec = StorageIO.load(self.run_id, "deployment-result.json")
         _config = EnvironmentDeploymentInfo.from_spec(
             self.environment_name, env_spec.get(self.environment_name), reader_type="remote"
         )
@@ -62,11 +63,10 @@ class AssetBasedLauncher:
             workflow.override_asset_based_launch_parameters(self._parameters)
 
         final_spec = workflow.dict(exclude_none=True, exclude_unset=True)
-
         cleaned_spec = self._cleanup_unsupported_properties(final_spec)
         run_data = service.submit_run(**cleaned_spec)
 
-        return run_data, None
+        return RunData(**run_data), None
 
     @staticmethod
     def _cleanup_unsupported_properties(spec: Dict[str, Any]) -> Dict[str, Any]: