+++ dbx/commands/execute.py
@@ -10,6 +10,7 @@ from dbx.api.context import RichExecutionContextClient
 from dbx.api.dependency.core_package import CorePackageManager
 from dbx.api.execute import ExecutionController
 from dbx.models.cli.execute import ExecuteParametersPayload
+from dbx.models.workflow.common.workflow_types import WorkflowType
 from dbx.options import (
     DEPLOYMENT_FILE_OPTION,
     ENVIRONMENT_OPTION,
@@ -42,7 +43,7 @@ def execute(
         None,
         "--task",
         help="""Task name (`task_key` field) inside the workflow to be executed.
-        Required if the workflow is a multitask job""",
+        Required if the workflow is a multitask job with more than one task""",
     ),
     deployment_file: Path = DEPLOYMENT_FILE_OPTION,
     requirements_file: Optional[Path] = REQUIREMENTS_FILE_OPTION,
@@ -94,6 +95,16 @@ def execute(
         config.build.no_build = True
 
     workflow = environment_config.payload.get_workflow(workflow_name)
+
+    if workflow.workflow_type == WorkflowType.pipeline:
+        raise Exception("DLT pipelines are not supported in the execute mode.")
+
+    if not task_name and workflow.workflow_type == WorkflowType.job_v2d1:
+        if len(workflow.task_names) == 1:
+            dbx_echo("Task key wasn't provided, automatically picking it since there is only one task in the workflow")
+            task_name = workflow.task_names[0]
+        else:
+            raise ValueError("Task key is not provided and there is more than one task in the workflow.")
 
     task: ExecuteTask = workflow.get_task(task_name) if task_name else workflow