+++ tests/unit/api/launch/test_functions.py
@@ -3,14 +3,16 @@ from unittest.mock import MagicMock
 import mlflow
 from pytest_mock import MockerFixture
 
-from dbx.api.launch.functions import cancel_run, load_dbx_file, wait_run
+from dbx.api.launch.functions import cancel_run, wait_run
+from dbx.api.launch.runners.base import RunData
+from dbx.api.storage.io import StorageIO
 from dbx.utils.json import JsonUtils
 
 
 def test_cancel(mocker: MockerFixture):
     wait_mock = mocker.patch("dbx.api.launch.functions.wait_run", MagicMock())
     client = MagicMock()
-    cancel_run(client, {"run_id": 1})
+    cancel_run(client, RunData(run_id=1))
     wait_mock.assert_called()
 
 
@@ -25,7 +27,7 @@ def test_wait_run(mocker: MockerFixture):
         ),
     )
     client = MagicMock()
-    wait_run(client, {"run_id": 1})
+    wait_run(client, RunData(**{"run_id": 1}))
 
 
 def test_load_file(tmp_path):
@@ -34,5 +36,5 @@ def test_load_file(tmp_path):
         _file = tmp_path / "conf.json"
         JsonUtils.write(_file, content)
         mlflow.log_artifact(str(_file.absolute()), ".dbx")
-        _result = load_dbx_file(test_run.info.run_id, "conf.json")
+        _result = StorageIO.load(test_run.info.run_id, "conf.json")
         assert _result == content