+++ tests/unit/commands/test_deploy_jinja_variables_file.py
@@ -9,19 +9,19 @@ from tests.unit.conftest import (
 )
 
 
-def test_incorrect_file_name(temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client):
+def test_incorrect_file_name(temp_project: Path, mlflow_file_uploader, mock_storage_io, mock_api_v2_client):
     deploy_result = invoke_cli_runner(["deploy", "--jinja-variables-file", "some-file.py"], expected_error=True)
     assert "Jinja variables file shall be provided" in str(deploy_result.exception)
 
 
-def test_non_existent_file(temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client):
+def test_non_existent_file(temp_project: Path, mlflow_file_uploader, mock_storage_io, mock_api_v2_client):
     deploy_result = invoke_cli_runner(
         ["deploy", "--jinja-variables-file", "some-non-existent.yml"], expected_error=True
     )
     assert "file is non-existent" in str(deploy_result.exception)
 
 
-def test_passed_with_unsupported(temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client):
+def test_passed_with_unsupported(temp_project: Path, mlflow_file_uploader, mock_storage_io, mock_api_v2_client):
     file_name = "jinja-template-variables-file.yaml"
     src_vars_file = get_path_with_relation_to_current_file(f"../deployment-configs/jinja-vars/{file_name}")
     dst_vars_file = Path("./conf") / file_name
@@ -33,7 +33,7 @@ def test_passed_with_unsupported(temp_project: Path, mlflow_file_uploader, mock_
     assert "deployment file is not based on Jinja" in str(deploy_result.exception)
 
 
-def test_passed_with_inplace(temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client):
+def test_passed_with_inplace(temp_project: Path, mlflow_file_uploader, mock_storage_io, mock_api_v2_client):
     invoke_cli_runner(["configure", "--enable-inplace-jinja-support"])
     file_name = "jinja-template-variables-file.yaml"
     src_vars_file = get_path_with_relation_to_current_file(f"../deployment-configs/jinja-vars/{file_name}")
@@ -43,7 +43,7 @@ def test_passed_with_inplace(temp_project: Path, mlflow_file_uploader, mock_dbx_
     assert isinstance(rdr, Jinja2ConfigReader)
 
 
-def test_jinja_vars_file_api(mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client, temp_project: Path):
+def test_jinja_vars_file_api(mlflow_file_uploader, mock_storage_io, mock_api_v2_client, temp_project: Path):
     jinja_vars_dir = get_path_with_relation_to_current_file("../deployment-configs/jinja-vars/")
     project_config_dir = temp_project / "conf"
     vars_file = Path("./conf/jinja-template-variables-file.yaml")
@@ -59,7 +59,7 @@ def test_jinja_vars_file_api(mlflow_file_uploader, mock_dbx_file_upload, mock_ap
         assert _def == definitions[0]
 
 
-def test_jinja_vars_file_cli(mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client, temp_project: Path):
+def test_jinja_vars_file_cli(mlflow_file_uploader, mock_storage_io, mock_api_v2_client, temp_project: Path):
     deployment_file_name = "09-jinja-with-custom-vars.yaml.j2"
     vars_file_name = "jinja-template-variables-file.yaml"
     project_config_dir = temp_project / "conf"